chore: rename from view to views
