+++ lib/presentation/restaurant_detail/views/restaurant_detail_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:restaurantour/core/di/di.dart';
 import 'package:restaurantour/domain/entities/entities.dart';
 import 'package:restaurantour/presentation/restaurant_detail/cubit/cubit.dart';
-import 'package:restaurantour/presentation/restaurant_detail/view/view.dart';
+import 'package:restaurantour/presentation/restaurant_detail/views/views.dart';
 
 class RestaurantDetailPage extends StatelessWidget {
   const RestaurantDetailPage({super.key, this.restaurantEntity});
