+++ lib/presentation/restaurant_list/views/restaurant_list_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:restaurantour/core/di/di.dart';
 import 'package:restaurantour/presentation/restaurant_detail/cubit/cubit.dart';
 import 'package:restaurantour/presentation/restaurant_list/bloc/bloc.dart';
-import 'package:restaurantour/presentation/restaurant_list/view/view.dart';
+import 'package:restaurantour/presentation/restaurant_list/views/views.dart';
 
 class RestaurantListPage extends StatelessWidget {
   const RestaurantListPage({super.key});
