+++ lib/presentation/restaurant_list/widgets/all_restaurants_list_widget.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:restaurantour/core/utils/utils.dart';
-import 'package:restaurantour/presentation/restaurant_detail/view/view.dart';
+import 'package:restaurantour/presentation/restaurant_detail/views/views.dart';
 import 'package:restaurantour/presentation/restaurant_list/bloc/bloc.dart';
 import 'package:restaurantour/presentation/widgets/widgets.dart';
