+++ lib/presentation/restaurant_list/widgets/favorite_restaurants_list_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:restaurantour/core/utils/utils.dart';
 import 'package:restaurantour/presentation/restaurant_detail/cubit/cubit.dart';
-import 'package:restaurantour/presentation/restaurant_detail/view/view.dart';
+import 'package:restaurantour/presentation/restaurant_detail/views/views.dart';
 import 'package:restaurantour/presentation/widgets/widgets.dart';
 
 class FavoriteRestaurantsListWidget extends StatefulWidget {
